feat: CardState 구현 및 적용
